Fix retrieving of CCYs
Remote all balances that are zero in Coinbase. This way the currency codes for such balances (atm. XRP and REP-V2) are not queried at.
besides this removes tons of tokens i do not own from my moneymoney coinbase balance. 

removed. signature as a am not able to compute it myself